+++ CHANGE_LOG.md
@@ -3,6 +3,8 @@
 # Conium event
 1. Renamed events, remove suffixes 'event' in the events list.
 2. Move more event trigger to intermediary.
+3. Fix bug of event 'ITEM_USE' and 'ITEM_USED' not imported in the common script
+4. Fix bug of event 'ITEM_USE' and 'ITEM_USED' not triggered when player using item stack
 
 ## New event support
 Add events support: